+++ NEWS.md
@@ -6,6 +6,7 @@
 * 32 palettes have been added from PrettyCols package. (#126)
 * 15 palettes have been added from waRhol package. (#131)
 * 18 palettes have been added from MexBrewer package. (#137)
+* 24 palettes have been added from ltc package. (#140)
 
 # paletteer 1.5.0